+++ src/Yaml/LibYaml.php
@@ -21,7 +21,7 @@ class LibYaml implements ParserInterface
      *
      * @throws ParseException
      *
-     * @return array
+     * @return null|array
      */
     public function parseFile($path)
     {