+++ src/Yaml/Sf2Yaml.php
@@ -22,7 +22,7 @@ class Sf2Yaml implements ParserInterface
      *
      * @throws ParseException
      *
-     * @return array
+     * @return null|array
      */
     public function parseFile($path)
     {