+++ CHANGELOG.md
@@ -8,6 +8,8 @@ The format is based on [Keep a Changelog] and Pipelines adheres to
 [Semantic Versioning]: https://semver.org/spec/v2.0.0.html
 
 ## [unreleased]
+### Change
+- html-docs: Upgrade mkdocs-material to 8.2.9
 
 ## [0.0.63] - 2022-04-14
 ### Change