+++ CHANGELOG.md
@@ -7,7 +7,7 @@ The format is based on [Keep a Changelog] and Pipelines adheres to
 [Keep a Changelog]: https://keepachangelog.com/en/1.0.0/
 [Semantic Versioning]: https://semver.org/spec/v2.0.0.html
 
-## [unreleased]
+## [0.0.64] - 2022-04-20
 ### Add
 - Support digests as an alternative to tags for container images in
   pipeline steps, see [Digest]
@@ -666,4 +666,5 @@ The format is based on [Keep a Changelog] and Pipelines adheres to
 [0.0.61]: https://github.com/ktomk/pipelines/releases/tag/0.0.61
 [0.0.62]: https://github.com/ktomk/pipelines/releases/tag/0.0.62
 [0.0.63]: https://github.com/ktomk/pipelines/releases/tag/0.0.63
+[0.0.64]: https://github.com/ktomk/pipelines/releases/tag/0.0.64
 [unreleased]: https://github.com/ktomk/pipelines